Add Markdown-specific sidebar transitions
For the transitions that include a sidebar, the transition from main body to
first sidebar involves placing a "</div>" tag, where the matching "<div>" tag
is part of the layout.  Because the opening "<div>" is part of the layout, and
not the body, Kramdown doesn't find it.  So, Kramdown assumes that this is
text, not HTML.

This implements a yucky workaround, which means these transitions need to be
Markdown-specific.  The yucky workaround tells Kramdown to _not_ process the
"</div>", but Kramdown still insists on putting it into a "<p>" block.  This
means we end up having an unmatched "</p>", but I'm not sure what else we can
do.