Add useless extra transition-to-right
See the previous commit.  We needed special Markdown versions of transition
includes, but this include does not need that special treatment.  However, to
make things cleaner for the user, we will create it.

We don't need special treatment for this file because the closing "</div>" tag
is matched by Kramdown with the opening "<div>" tag in the
transition-to-left.md file.