Move bottom-left content into includes
The entire bottom-left of the page (which, right now, is just the social media
block) was hard-coded into the template.  This change splits out the
bottom-left content into an include, and it also moved the "Connect With Us"
block into a separate include.  This gives users the opportunity to customize
it.