Add GridFTP security/logging configuration
For Medium Risk, we need TLS version 1.1 or later, and we need to
"Enable any available application logging that would assist in a
forensic investigation in the event of a compromise."  Both of these
things can be done in GridFTP configuration, so instructions are added
to enable both.

We have to put the instructions here, because
/var/lib/globus-connect-server only gets created when
globus-connect-server-setup is run.

Updates #25 and #27