Add MyProxy OAuth CSS and logo
MyProy OAuth has its own CSS; when you add a site-specific CSS, both CSS files
end up being loaded (ours is loaded second).  This means that our CSS will not
only have to contain the stuff we want, it will have to be sure to override the
general CSS things that we want to change (even if that means setting a
property to `inherit`).

The logo is downloaded by the server admin, from the Identity web site.