Update Client instructions for new GCP
The new Globus Connect Personal client brings with it a new, cleaner
installation process.  The collection creation process is now much
improved, and is largely automated.  Also, Globus have written their own
set of guides to walk users through installation.

Because of the above, much of the client installation page is gone,
replaced with links to Globus' documentation.  What remains are some of
the endpoint customization instructions, as well as the section about
adding allowed paths (though this largely duplicates Globus' docs, it is
good to repeat).