Add Gemfile.lock to the repo
This removes Gemfile.lock from .gitignore, and adds it to the repo.

This is being done primarily for Dependabot, which needs a Gemfile.lock
in order to check for dependency problems.  (Let's ignore that the
Gemfile is only being used during development to host a local Jekyll
server...)

Anyway, this does have one advantage: When multiple people are working
on the site, running the same Jekyll means everyone should (hopefully)
see the same render of content!