policies: Add language for cloud providers
The "What Is Provided?" section is updated to mention the conenctors we
have access to.

The "Data Collected & Privacy" section is updated to note that, if a
cloud provider is involved in the transfer, they (either the
accountholder or the provider) may keep logs.

The "Other Policies & Agreements" section adds references to the terms &
conditions for AWS, Box, Google, and Wasabi.  Each reference notes that
the public terms may be modified by an agreement between the cloud
provider and Stanford.