Add a cloud page
This (finally!) adds a page on the cloud add-ons that we can use.

At the top is the list of (most of) the cloud add-ons available.  Some
premium storage connectors, like HPSS, aren't listed as they are very
special-case and not of interest to the vast majority of readers.

For each connector not currently licensed (that is, Ceph, iRODS,
Microsoft Azure Blob Storage, and Microsoft OneDrive), a boilerplate note is
included, with a suggestion to contact us.

Google Cloud Storage and Wasabi have relatively little information.
I'll need to actually do some work with the specific platform, in order
to provide better documentation.

Stanford Box and Google Drive are essentially complete.  Medicine Box
has a mention in the Stanford Box section, about how it can't be used.

Amazon S3 has the most detail.  Since the permissions required are
fairly complex, a separate page will need to be created.