cloud/s3: Add cross-account access instructions
This explains how to set up cross-account access for am IAM User, such
that Globus can access a cross-account bucket.

Unlike the rest of the page, this section is targeted at someone who
does not use Globus.  That being said, there is a mention at the top for
Globus users, noting that instead of setting up cross-account access, it
might be possible to set up a guest collection instead.