cloud: Add Google Drive instructions
This adds detailed instruction for Google Drive.

There are a number of limitations, both Globus-specific limitations and
Drive limitations.  The Drive limitations are not specific to Globus,
but I felt it was best to include them here anyway.

A special callout is made to the 'files' that Google products store in
Drive.  They can be downloaded, but what ends up being downloaded is
weird.

Finally, include instructions for setup and access, using the same
structure that I used in the S3 instructions.