scenarios: Add "How do I get started?"
I realized that there is now 'quick start' for Globus @ Stanford; this
is the answer to that need.

The answer starts by noting that you need Globus running at each end of
the transfer, so it assumes that you want to transfer some data
somewhere.

The first action to perform is creating a Globus account.  An optional
challenge is provided, to link your ORCID to your Stanford Globus
identity.

Next, information is provided on how to locate a collection (for
servers) or endpoint (for endpoints).  For those who want to install on
a server, a link to the Globus GCS instructions is provided.  For those
who want to install on an endpoint, a link is provided to the GCP
instructions.  A link is also provided to the list of supported cloud
providers.

Finally, a link is provided to the Globus instructions on how to do a
basic transfer.  That should be enough to get people going!