Retire Box content
With Box being unavailable, and it being pretty clear that it's not
coming back anytime soon, it's time to retire the Box content.

The cloud page is updated with a note of retirement, and the link to the
Box-specific page is removed.

The Box-specific page is updated with note of retirement, but the page
itself is _not_ deleted, as it's been indexed in search engines, and it
would be better for people to come to a page saying "It's been retired",
as opposed to a broken link.