Add post for GCSv4 deprecation
This repeats the GCSv4 deprecation announcement: Security updates only
starting August 1, with GCSv4 completely disabled on December 18.

The post includes a note that GCP is unaffected, though users should
make sure they're on 3.2.0 or later (due to the Globus CA change).

The docs in the server directory will be marked deprecated in a
different commit.

A **big** plug is made for GCSv5 migration, and what to expect.  Notes
are also provided on what firewall changes are needed, and how to get
support.