policies: Add Dropbox
Dropbox is added to the list of connectors we can access.  Also, add a
link to Dropbox's Terms of Service, with a note about it being modified:
Although UIT doesn't provide Dropbox, it is a known vendor in Oracle
Financials, and it looks like there's at least one Contract, so include
the standard terms-as-modified text.