README: Clarify OS for instructions, and add Git
This updates the README to clarify that the last two steps are written
for macOS and Linux.  The Linux steps have also been updated to explain
what packages to use for each distro.

The README also assumed you knew how to use Git!  Now, the README has
you install the GitHub CLI, and explains how to use that to clone the
repository.