Update README with instructions for Windows
This updates the README to include instructions on how to set up the dev
environment on Windows.

The biggest part comes in the installation instructions for Windows:
Besides Ruby, it also covers the RubyInstaller (to bring in a build
environment), Git for Windows, and the GitHub CLI.  Bundler needs some
configuration, but otherwise it works the same as on Linux.  Though that
is largely thanks to both the `gem update --system` and the change in
the Bundler version used in the lockfile; those had to be updated to fix
issues around path handling (older versions had bugs when paths
contained spaces).

Fetching the repo and running Jekyll are essentially the same across all
releases, but this commit also updates some of those instructions,
particularly for the first run of the GitHub CLI.