Archive the GCSv4-to-5 Transition page
At the end of last year, Globus HQ dropped support for GCSv4.  GCSv4
installations stopped working, so it's no longer possible to migrate
from GCSv4 to GCSv5.

This adds a note to the transition page, and removes it from the ToC.
The page remains, so that it can be found by folks who have the link.