server5: Big updates to the Install page
This makes the Install page usable.  It starts by removing all the parts
about getting an Endpoint ID, since this is now handled automatically.

The Installation process is updated to point users either towards the
quickstart or towards the full installation guide.  A note is added to
let users know when to reach out to us (for making an endpoint managed).