S3:
This makes a number of changes to the S3 collections that we advertise.

The region-specific collections have been removed: us-west-2 has
been promoted to be the 'all regions' collection, and the other
collections are no longer mentioned.

(Fun fact: The collections for us-west-1 and us-east-2 were never
accessed, and so have been deleted immediately.)

Also, on the Cloud page, the SCG S3 collection is no longer mentioned,
and is replaced with a link to the 'all regions' collection (that used
to be us-west-2).

(Another fun fact: The 'all regions' collection is still pointing to
us-west-2, but the S3 connector is smart enough to look up what region a
bucket is associated with, and use that region's S3 endpoint.  So, there
will always be a little bit of traffic going to us-west-2, but actual
data transfers will go direct to the appropriate region.)

Finally, a new region has been added for requester-pays buckets.
Nothing else is said about those buckets right now; folks are generally
told when a bucket they are accessing is requester-pays.