cloud: Google Cloud works with base SUNetIDs now
Google Cloud used to require a full (or full-sponsored) SUNetID.  Folks
with a base SUNetID would either not have a Google account, or it would
be disabled.  Now, it's possible for folks with a base SUNetID to have a
Google account, using a mechanism similar to sponsorship.

This changes the Google Cloud callout box language appropriately.

For Google Drive, a full (or full-sponsored) SUNetID is still required.
So, the wording for Google Drive has also been adjusted.