Switch info-box captures to h4
Info-boxes normally get their contents as parameters, with the HTML
being generated via `_includes/info-box.html`.  But in some places, we
provide raw HTML to the info-box, via a `{% capture %}` and the info-box
`raw=` parameter.

In the original template, info-box headers used an `<h2>`.  Now, they
use a `<h4>`.  So, the raw HTML needs to be changed to match.