cloud: Slim down remaining mentions of Box
Since Stanford Box has been retired for a while, this change slims down
the Box content on the Cloud overview page: The "Stanford Box" entry is
removed (though "Medicine Box" remains), and the Box section now simply
says that UIT doesn't provide Box as a service.  The Medicine Box
mention remains, though it's been moved from an info-box to being part
of the normal body text.

Finally, Box is removed from the sidebar ToC.  The original Box page
remains on the site, even though it is not linked from anywhere.  This
is so that folks who still have the link can find it.  The original Box
page does have notations on it, talking about the Box retirement.