The start of the new Elm page.
The S3 page was used as a base, and then heavily rewritten for Elm.  The
limitations are different, and the process for getting an IAM User is
different.  The rest of the process is largely the same as for S3, but I
decided to keep all of the docs, limitations, etc. on this page, instead
of referencing the AWS S3 page.

One big TODO: For the Customer User Policy, this should be generated for
the user using JavaScript, which asks for the bucket name(s), and if
write and/or delete are OK.