Workflow: Run Pages build on all branches
Instead of only doing a GitHub Pages build on the main branch, this will
run a GitHub Pages build on all branches.  But, only the main branch's
build will be pushed live.

Also, this changes how long the GitHub Pages build result artifact will
be retained.  By default, the artifact is kept for one day, since it is
normally just needed for the deployment action.  Now, the artifacts will
be kept for a week.  That is useful for branches; we can download the
artifact and see what the built site would look like.