Workflow: Cache live GitHub Pages site
Thinking ahead, it may be useful to compare a pending change to the
live site.  In order to do that, we need a way to access the live site.
This can be a problem since GitHub Actions artifacts have a lifetime.

Following the advice in [1], use the GitHub Actions cache to store a
copy of the live site during deployment.

[1]: https://github.com/actions/upload-pages-artifact/issues/57#issuecomment-1833968666