Elm: Reorder & expand Setting a User Policy
In the not-yet-linked-anywhere Elm page, change the instructions around
to make it clear that you get a default policy to start, you want to
delete that policy, and then you want to copy-paste a new policy (using
the template policy provided).  This changes one screenshot, and adds a
few more.

This is all coming from feedback registered by Chris Haley.  Thanks,
Chris!

Reported-By: Christopher Haley <clhaley@stanford.edu>