Elm: Change S3 wildcard away from `*`
Per Globus' documentation[0], the `s3:ListAllMyBuckets` Action must be
allowed for Resource `*`.  But in the MinIO version that Elm is using,
that throws an error.  So instead, switch to the fully-qualified
wildcard of `arn:aws:s3:::*`.

The root cause is a lack of MinIO support for the `*` wildcard.  This
was fixed upstream in [1], but Elm has not yet moved to one of the
releases that includes the commit.

[0]: https://docs.globus.org/premium-storage-connectors/v5.4/aws-s3/#permissions-anchor

[1]: https://github.com/minio/pkg/commit/f99581f2bdb29f780eef1401606bbf1933ee4e69,