cloud: Add an entry for OCI Object Storage
Globus has recently validated that Oracle Cloud Infrastructure (OCI) Object
Storage works with Globus, using the S3 Connector.  See the news post at https://www.globus.org/news/globus-adds-support-for-oracle-cloud-infrastructure-oci-object-storage

This adds an item to the Cloud page, noting that OCI access is included
in our subscription, but that we do not provide access to it ourselves.